requirements.txt
# Core deep learning and data processing
tensorflow>=2.13.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0

# Data visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Data download and manipulation
gdown>=4.6.0
pandas>=1.3.0

# Optional but recommended
# jupyterlab>=3.0.0  # For interactive development
# tensorboard>=2.8.0  # For training visualization